@@ -5,11 +5,11 @@ apply plugin: 'com.android.library'
 apply plugin: 'maven'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
 
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 26
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
     }
@@ -29,7 +29,7 @@ dependencies {
     api project(':lib')
     api 'io.reactivex:rxandroid:1.2.1'
     api 'io.reactivex:rxjava:1.1.10'
-    api "com.android.support:support-v4:$support_lib_version"
+    api 'androidx.legacy:legacy-support-v4:1.0.0'
 }
 
 task sourcesJar(type: Jar) {
