@@ -1,8 +1,8 @@
 package jp.programminglife.libpljp.android.uuidgenerator;
 
-import android.databinding.DataBindingUtil;
+import androidx.databinding.DataBindingUtil;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.text.Editable;
 import android.view.View;
 import android.view.View.OnClickListener;
